backend/requirements.txt
fastapi
uvicorn
openai-whisper
yt-dlp
opencv-python
ultralytics
easyocr
langchain
langchain-huggingface
langchain-community
ollama
faiss-gpu
panns-inference
torchlibrosa
soundfile
NudeNetClassifier
deepface